requirements.txt
Flask
Flask-Cors
python-dotenv
pandas
scikit-learn
googletrans==3.1.0a0
streamlit
requests
matplotlib
pandas
SpeechRecognition
